requirements.txt
python-telegram-bot==13.15
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0